+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add a new `wanted` game state in the web UI, replaces `incomplete` which now means games with some (but not all) its ROMs
 - Fix a regression when importing multi-track CHDs which contain tracks identical to other games
 - Fix parsing archives containing files with `=` symbols in their names
+- Fix several issues affecting the import of MAME DATs
 
 # 0.20.1